+++ R/CreateOutputs.R
@@ -1,196 +1,199 @@
 #' @import ggplot2
 
-# upp = User's Prior Projection
-
-GetModelName <- function(dt) {
-  model.name <- dt[, factor(paste0(as.integer(hasE), as.integer(hospInf), as.integer(hospRate)))]
-  return(model.name)
-}
-
-GetPlotTitle <- function(posterior.niter) {
-  if (posterior.niter == 0) {
-    warn.str <- "\nNo credibility interval available"
-  } else if (posterior.niter < 100) {
-    warn.str <- "\nVery low number of iterations, do not use for inference"
-  } else if (posterior.niter < 500) {
-    warn.str <- "\nLow number of iterations, use with caution"
-  } else {
-    warn.str <- ""
-  }
-  paste0("Posterior Distribution, niter = ", posterior.niter, warn.str)
-}
-
-GetExcelOutput <- function(sim, upp, in.bounds, upp.in.bounds, date.range, filestr, all.inputs.str) {
-  output.list <- list(hosp = NULL, icu = NULL, vent = NULL, active.cases = NULL, total.cases = NULL)
-  output.names <- names(output.list)
-  
-  probs2 <- c(0.95, 1, 0.15, 0.25, seq(0.55, 0.9, by = 0.05))
-  for (j in output.names) {
-    sim.accepted <- sim[[j]][, in.bounds, drop = F]
-    if (is.null(sim)) {
-      quant1 <- quant2 <- NULL
-    } else {
-      quant1 <- rowQuantiles(sim.accepted, probs = c(0, 0.05, 0.5))
-      quant2 <- rowQuantiles(sim.accepted, probs = probs2)
-    }
-    output <- data.table(date = date.range, quant1, upp = upp[[j]], quant2)
-    output[, notes := ""]
-    output[1, notes := GetPlotTitle(posterior.niter = sum(in.bounds))]
-    output[2, notes := paste0("User's Prior Projection ", ifelse(upp.in.bounds, "accepted", "rejected"))]
-    output.list[[j]] <- cbind(output.list[[j]], output)
-  }
-  output.list$all.inputs = all.inputs.str
-  filestr.out <- paste0(filestr, ".xlsx")
+GetExcelOutput <- function(quantile.list, inputs) {
+  options("openxlsx.numFmt" = "0.0")
+  display.date.range <- as.character(seq(inputs$model.inputs$start.display.date, inputs$model.inputs$end.date, by = "day"))
+  output.list <- lapply(quantile.list, function (quant) {
+    output <- data.table(date = display.date.range, quant[display.date.range, ])
+    return(output)
+  })
+  output.list$all.inputs = inputs$all.inputs.str
+  filestr.out <- paste0(inputs$internal.args$output.filestr, ".xlsx")
   openxlsx::write.xlsx(output.list, file = filestr.out)
   cat("\nExcel output: ", filestr.out, "\n")
   return(output.list)
 }
 
-PlotHist <- function(x, posterior.title, cap, xlab, in.bounds) {
-  if (uniqueN(x) > 1) {
-    g.list <- list()
-    for (prior in c(T, F)) {
-      if (prior) {
-        d <- data.table(x)
-        breaks <- NULL
-        title1 <- paste0("Prior Distribution, niter = ", length(x))
-      } else {
-        d <- data.table(x = x[in.bounds])
-        breaks <- unlist(ggplot_build(g.list[[1]])$data[[1]][, c("xmin", 'xmax')])
-        breaks <- sort(unique(round(breaks, digits = 8))) #round to fix problems with nearly identical breaks
-        title1 <- posterior.title
-      }
-      g <- ggplot(d, aes(x=x))
-      if (class(x) %in%c("logical", "integer", "Date", "factor") || all(x == round(x))) {
-        g <- g + geom_bar()
-      } else {
-        g <- g + geom_histogram(aes(y=..density..),     
-                                #   binwidth=.1,
-                                bins = 20,
-                                breaks = breaks,
-                                color="black", fill="white") +
-          geom_density(alpha=.2, fill="steelblue3") 
-      }
-      g <- g + labs(title = title1, caption = cap) + xlab(xlab)
-      g.list[[length(g.list) + 1]] <- g
-    }
-  } else {
-    g.list <- NULL
-    cat("No histogram for ", xlab, ", constant value = ", as.character(unique(x)), "\n", sep = "")
+#modification of rstan::stan_hist to deal with dates (also rstan::stan_hist crashes when using ExtendSim)
+StanHist <- function (object, pars, base.date, ...) {
+  dots <- rstan:::.add_aesthetics(list(...), c("fill", "color"))
+  samp <- melt(as.data.table(as.matrix(object, pars = pars)), id.vars = integer(0))
+
+  if (!is.null(base.date)) {
+    samp$value <- base.date + samp$value
   }
-  return(g.list)
+  thm <- rstan:::rstanvis_hist_theme()
+  base <- ggplot2::ggplot(samp, ggplot2::aes_string(x = "value", y = "..density.."))
+  graph <- base + do.call(ggplot2::geom_histogram, dots) +
+    ggplot2::xlab("") + thm
+  if (uniqueN(samp$variable) == 1) {
+    graph + ggplot2::xlab(unique(samp$variable))
+  } else {
+    graph + ggplot2::facet_wrap(~variable, scales = "free")
+  }
 }
 
-GetProjectionPlot <- function(short.term, niter, hosp.quantiles, bounds.without.multiplier, bounds.labels, plot.observed.data) {
+PlotHist <- function(fit, pars, base.date) {
+  base.date <- if (pars == "t_inter") base.date else NULL
+  return(StanHist(fit, pars = pars, fill="steelblue3", alpha = 0.2, bins=30, base.date = base.date))
+}
+
+GetYLabel <- function(data.type, long.name) {
+  switch(data.type,
+         hosp = "Number of COVID19 Patients in Hospital",
+         icu = "Number of COVID19 Patients in ICU",
+         deaths = "Number of COVID19 Deaths",
+         cum.admits = "Number of Cumulative COVID19 Admissions to Hospital",
+         stop("unexpected data.type"))
+}
+
+GetTitleLabel <- function(data.type) {
+  switch(data.type,
+         hosp = "Hospitalization",
+         icu = "ICU",
+         deaths = "Death",
+         cum.admits = "Cumulative Admission",
+         stop("unexpected data.type"))
+}
+
+GetProjectionPlot <- function(short.term, quantiles, data.type, inputs) {
+  obs.data <- inputs$obs.data[, .(date, conf = get(paste0(data.type, ".conf")), pui = get(paste0(data.type, ".pui")))]
+  if (all(is.na(obs.data$conf))) return(NULL)
+  quantiles.dt <- data.table(date = as.Date(rownames(quantiles[[data.type]])), quantiles[[data.type]])
+  dt.plot <- merge(obs.data, quantiles.dt, all = T, by = "date")
+
   if (short.term) {
     date.breaks <- "1 week"
-    max.date <- bounds.without.multiplier[!is.na(lower), max(date)] + 3
-    title1 <- "Short Term Hospitalization Projections"
+    max.date <- obs.data[!is.na(conf), max(date)] + 3
+    title1 <- paste("Short Term", GetTitleLabel(data.type), "Projection")
+    plot.observed.data <- inputs$internal.args$plot.observed.data.short.term
   } else {
     date.breaks <- "1 month"
-    max.date <- max(hosp.quantiles$date) - 3 #makes it look a little nicer when ending on first of the month
-    title1 <- "Long Term Hospitalization Projections"
+    max.date <- max(dt.plot$date) - 3 #makes it look a little nicer when ending on first of the month
+    title1 <- paste("Long Term", GetTitleLabel(data.type), "Projection")
+    plot.observed.data <- inputs$internal.args$plot.observed.data.long.term
   }
-  
-  if (niter < 500) {
-    subtitle1 <- GetPlotTitle(niter)
-  } else {
-    subtitle1 <- NULL
-  }
-  
-  plot.cred.int <- niter > 1
-  plot.upp <- niter < 100
- 
-  min.date <- min(bounds.without.multiplier$date) - 7 
-  obs.size <- 3
-  lb <- bounds.labels[1]
-  ub <- bounds.labels[2]
-  
-  dt.plot <- merge(hosp.quantiles, bounds.without.multiplier, all.x = T, by = "date")
-  dt.plot <- dt.plot[date >= min.date & date <= max.date]
-  gg <- ggplot(dt.plot, aes(x=date)) +    
-    theme_light()
-  
-  if (niter > 0) {
-    gg <- gg + geom_line(aes(y = `50%`, color = "Median"))
-  }
-  
+
   if (plot.observed.data) {
-    gg <- gg + 
-      geom_point(aes(y=upper, color = ub), size = obs.size, shape = 4, na.rm = T) +
-      geom_point(aes(y=lower, color = lb), size = obs.size, shape = 4, na.rm = T)
-    over.aes <- list(shape = c(if (niter > 0) NA, 4, 4), linetype = c(if (niter > 0) 1, 0, 0))
+    min.date <- obs.data[!is.na(conf), min(date)] - 7
+  } else {
+    min.date <- inputs$model.inputs$start.display.date - 7
+  }
+
+  obs.size <- 1.5
+  frac.pui <- inputs$frac_pui[name == data.type, mu]
+  lb <- "Confirmed"
+  ub <- paste0("Confirmed + ", 100 * frac.pui, "%PUI")
+  dt.plot[, upper := conf + frac.pui * pui]
+
+  dt.plot <- dt.plot[date >= min.date & date <= max.date]
+  gg <- ggplot(dt.plot, aes(x=date)) +
+    theme_light() +
+    geom_line(aes(y = `50%`, color = "Median"))
+
+
+  if (plot.observed.data) {
+    gg <- gg + geom_point(aes(y=conf, color = lb), size = obs.size, shape = 4, na.rm = T)
+    if (!all(is.na(dt.plot$upper))) {
+      gg <- gg + geom_point(aes(y=upper, color = ub), size = obs.size, shape = 4, na.rm = T)
+      over.aes <- list(shape = c(NA, 4, 4), linetype = c(1, 0, 0))
+    } else {
+      over.aes <- list(shape = c(NA, 4), linetype = c(1, 0))
+    }
   } else {
     over.aes <- list()
   }
-  
-  if (plot.cred.int) {
-    gg <- gg + geom_ribbon(aes(ymin=`25%`, ymax=`75%`, alpha = "25%-75%"), fill = "blue") +
-      geom_ribbon(aes(ymin=`15%`, ymax=`85%`, alpha = "15%-85%"), fill = "blue") +
-      geom_ribbon(aes(ymin=`5%`, ymax=`95%`, alpha = "5%-95%"), fill = "blue") 
-  }
-  
-  if (plot.upp) {
-    gg <- gg + geom_line(aes(y = upp), color = "yellow1", size = 1, linetype = "dashed")  + 
-      theme(panel.background = element_rect(fill = "grey80")) + 
-      annotate("text", x = median(dt.plot$date), y = 10, label = "Yellow line is User's Prior Projection (column E of input)")
-  } 
-  
-  gg <- gg + 
-    xlab("") + 
-    ylab("Number of COVID19 Patients in Hospital") +
-    labs(title = title1, subtitle = subtitle1) + 
+
+  gg <- gg + geom_ribbon(aes(ymin=`25%`, ymax=`75%`, alpha = "25%-75%"), fill = "blue") +
+    geom_ribbon(aes(ymin=`15%`, ymax=`85%`, alpha = "15%-85%"), fill = "blue") +
+    geom_ribbon(aes(ymin=`5%`, ymax=`95%`, alpha = "5%-95%"), fill = "blue")
+
+  gg <- gg +
+    xlab("") +
+    ylab(GetYLabel(data.type)) +
+    labs(title = title1) +
     scale_color_manual("", values = c("blue", "palegreen4", "red4"), breaks = c("Median", lb, ub)) +
     scale_alpha_manual("", values = c(0.2, 0.3, 0.4), breaks = c("5%-95%", "15%-85%", "25%-75%")) +
-    theme(plot.title = element_text(hjust = 0.5), plot.subtitle = element_text(hjust = 0.5)) + 
+    theme(plot.title = element_text(hjust = 0.5), plot.subtitle = element_text(hjust = 0.5)) +
     scale_x_date(date_breaks = date.breaks, date_labels = "%b %d", expand = expansion()) +
     guides(color = guide_legend("", order = 1, override.aes = over.aes), alpha = guide_legend("", order = 2)) +
     theme(legend.position = "bottom", plot.margin = margin(0.1, 0.2, 0, 0.1, "in"),
           axis.text.x=element_text(hjust = 0.7))
+  print(gg)
   return(gg)
 }
 
-GetPdfOutput <- function(hosp, in.bounds, all.params, filestr, bounds.without.multiplier, internal.args) {
-  post.niter <- sum(in.bounds)
-  posterior.title <- GetPlotTitle(post.niter)
-  filestr.out <- paste0(filestr, ".pdf")
+GetRtPlot <- function(quantiles, inputs) {
+  sim.dates <- inputs$internal.args$simulation.start.date + 1:nrow(quantiles)
+  min.date <- as.Date("2020/4/1") #inputs$model.inputs$start.display.date - 7
+  dt.plot <- data.table(date = sim.dates, quantiles)[date > min.date]
+
+  gg <- ggplot(dt.plot, aes(x=date)) +
+    theme_light() +
+    geom_line(aes(y = `50%`, color = "Median"))
+  gg <- gg + geom_ribbon(aes(ymin=`25%`, ymax=`75%`, alpha = "25%-75%"), fill = "blue") +
+    geom_ribbon(aes(ymin=`15%`, ymax=`85%`, alpha = "15%-85%"), fill = "blue") +
+    geom_ribbon(aes(ymin=`5%`, ymax=`95%`, alpha = "5%-95%"), fill = "blue")
+
+  gg <- gg +
+    xlab("") +
+    ylab("Re") +
+    labs(title = "Effective Reproduction Number") +
+    scale_color_manual("", values = "blue") +
+    scale_alpha_manual("", values = c(0.2, 0.3, 0.4), breaks = c("5%-95%", "15%-85%", "25%-75%")) +
+    theme(plot.title = element_text(hjust = 0.5), plot.subtitle = element_text(hjust = 0.5)) +
+    scale_x_date(date_breaks = "1 week", date_labels = "%b %d", expand = expansion()) +
+    guides(color = guide_legend("", order = 1), alpha = guide_legend("", order = 2)) +
+    theme(legend.position = "bottom", plot.margin = margin(0.1, 0.2, 0, 0.1, "in"),
+          axis.text.x=element_text(hjust = 0.7)) +
+    geom_hline(yintercept = 1, lty = "dashed")
+  print(gg)
+  return(gg)
+}
+
+
+GetPdfOutput <- function(fit, quantiles, inputs) {
+  devlist <- grDevices::dev.list()
+  sapply(devlist[names(devlist) == "pdf"], grDevices::dev.off) #shuts down any old pdf (if there was a crash part way)
+
+  filestr.out <- paste0(inputs$internal.args$output.filestr, ".pdf")
   grDevices::pdf(file = filestr.out, width = 9.350, height = 7.225)
-  
-  bounds.labels <- c(internal.args$lower.bound.label, internal.args$upper.bound.label)
-  print(short.term <- GetProjectionPlot(short.term = T, niter = post.niter, hosp.quantiles = hosp, bounds.without.multiplier = bounds.without.multiplier, bounds.labels = bounds.labels, plot.observed.data = internal.args$plot.observed.data.short.term))
-  print(long.term <- GetProjectionPlot(short.term = F, niter = post.niter, hosp.quantiles = hosp, bounds.without.multiplier = bounds.without.multiplier, bounds.labels = bounds.labels, plot.observed.data = internal.args$plot.observed.data.long.term))
-  
-  intervention.multiplier.str <- grep("^intervention[[:digit:]]+\\.multiplier$", names(all.params), value = T)
-  captions <- list()
-  for (int.num in 0:length(intervention.multiplier.str)) {
-    cumRe <- all.params[, r0.initial]
-    cumRe.name <- paste0(".Re.", int.num)
-    captions[[cumRe.name]] <- paste0(cumRe.name, " = r0.initial")
-    for (i in seq_len(int.num)) {
-      cur.multiplier <- all.params[[paste0("intervention", i, ".multiplier")]]
-      stopifnot(!is.null(cur.multiplier))
-      cumRe <- cumRe * cur.multiplier
-      captions[[cumRe.name]] <- paste0(captions[[cumRe.name]], " * intervention", i, ".multiplier")
-    }
-    all.params[[cumRe.name]] <- cumRe
+
+  short.term <- long.term <- list()
+  for (i in names(quantiles)) {
+    short.term[[i]] <- GetProjectionPlot(short.term = T, quantiles = quantiles, data.type = i, inputs = inputs)
+    long.term[[i]] <- GetProjectionPlot(short.term = F, quantiles = quantiles, data.type = i, inputs = inputs)
   }
-  suppressWarnings(all.params[, model := factor(paste0(as.integer(latent.period > 0), as.integer(patients.in.hosp.are.infectious), as.integer(use.hosp.rate)))]) #suppress "Invalid .internal.selfref detected and fixed by ..." (not important)
-  captions$model <- "(HasE  InfectInHosp   RateToHosp)"
-  if (post.niter >= 1) {
-    for (param.name in sort(names(all.params[,-"r0.initial"]))) {
-      g.list <- PlotHist(all.params[[param.name]], posterior.title, captions[[param.name]], param.name, in.bounds)
-      sapply(g.list, print)
-    }
-  }
+
+  rt.date <- max(inputs$obs.data$date) - 14
+
+  rt.all <- rstan::extract(fit, pars = "Rt")[[1]]
+  rt.quantiles <- colQuantiles(rt.all, probs = c(0.05, 0.15, 0.25, 0.5, 0.75, 0.85, 0.95))
+  GetRtPlot(rt.quantiles, inputs)
+
+  date.index <- as.numeric(rt.date - inputs$internal.args$simulation.start.date)
+  rt.pars <- paste0("Rt[", date.index, "]")
+  rt <- rstan::extract(fit, pars = rt.pars)[[1]]
+  rt.quantiles <- round(quantile(rt, c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95)), 2)
+  print(rt.quantiles)
+  subtitl <- paste0(capture.output(print(rt.quantiles)), collapse = "\n")
+  g <- PlotHist(fit, rt.pars) + xlab(NULL) + labs(title = paste("Rt as of", rt.date), subtitle = subtitl) + theme(plot.subtitle=element_text(family="Courier"))
+  print(g)
+
+  pars <- c("r0", "duration_latent", "duration_rec_mild", "duration_pre_hosp", "duration_hosp_mod",
+            "duration_hosp_icu", "frac_hosp", "frac_icu", "frac_mort",
+            "beta_multiplier", "t_inter", "len_inter", "frac_PUI", "ini_exposed")
+  lapply(pars, function (p) print(PlotHist(fit, p, base.date = inputs$internal.args$simulation.start.date)))
+
   grDevices::dev.off()
   cat("\nPDF output: ", filestr.out, "\n")
   return(list(short.term = short.term, long.term = long.term))
 }
 
 TestOutputFile <- function(filestr) {
-  z <- file.create(filestr) #filestr has no extension, it's not the actual .pdf or .xlsx output
-  if (!z) {
+  success <- file.create(filestr) #filestr has no extension, it's not the actual .pdf or .xlsx output
+  if (!success) {
     stop("Unable to write to output file: ", filestr, "\nCheck your working directory and the value of internal.args$output.filestr")
   }
-  file.remove(filestr) 
+  file.remove(filestr)
 }